refactor to entity component system architecture
