+++ core/src/io/alfami/builders/WorldBuilder.java
@@ -3,27 +3,27 @@ package io.alfami.builders;
 import com.badlogic.ashley.core.Engine;
 import com.badlogic.ashley.core.Entity;
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Animation;
+import com.badlogic.gdx.graphics.g2d.Animation.PlayMode;
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
-import com.badlogic.gdx.graphics.g2d.Animation.PlayMode;
 import com.badlogic.gdx.maps.MapLayer;
 import com.badlogic.gdx.maps.MapLayers;
 import com.badlogic.gdx.maps.MapObject;
 import com.badlogic.gdx.maps.objects.RectangleMapObject;
 import com.badlogic.gdx.maps.tiled.TiledMap;
-import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.physics.box2d.Body;
 import com.badlogic.gdx.physics.box2d.BodyDef;
 import com.badlogic.gdx.physics.box2d.FixtureDef;
 import com.badlogic.gdx.physics.box2d.PolygonShape;
 import com.badlogic.gdx.physics.box2d.World;
+import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.utils.Array;
 
 import io.alfami.components.AnimationComponent;
+import io.alfami.components.HealthBarComponent;
 import io.alfami.components.HealthComponent;
 import io.alfami.components.MovementComponent;
 import io.alfami.components.PlayerComponent;
@@ -37,33 +37,28 @@ public class WorldBuilder {
 	private final TiledMap tiledMap;
 	private final World world;
 	private final Engine engine;
-
-	private final AssetManager assetManager;
 	private final TextureAtlas actorAtlas;
+	private Stage stage;
 
-	public WorldBuilder(TiledMap tiledMap, World world, Engine engine) {
+	public WorldBuilder(TiledMap tiledMap, World world, Engine engine, Stage stage) {
 		this.tiledMap = tiledMap;
 		this.world = world;
 		this.engine = engine;
-		assetManager = GameManager.instance.assetManager;
+		this.stage = stage;
 		actorAtlas = new TextureAtlas(Gdx.files.internal("hero/hero.atlas"));
 	}
 
 	public void buildAll() {
-		buildMap();
+
 	}
 
-	private void buildMap() {
+	public void buildMap() {
 		MapLayers mapLayers = tiledMap.getLayers();
 
-		int mapWidth = ((TiledMapTileLayer) mapLayers.get(0)).getWidth();
-		int mapHeight = ((TiledMapTileLayer) mapLayers.get(0)).getHeight();
 		MapLayer wallLayer = mapLayers.get("collisions"); // wall layer
 
 		for (MapObject mapObject : wallLayer.getObjects()) {
 			Rectangle rectangle = ((RectangleMapObject) mapObject).getRectangle();
-
-//			correctRectangle(rectangle);
 
 			BodyDef bodyDef = new BodyDef();
 			bodyDef.type = BodyDef.BodyType.StaticBody;
@@ -79,11 +74,10 @@ public class WorldBuilder {
 			body.createFixture(fixtureDef);
 			polygonShape.dispose();
 		}
-		createPlayer(752, 320);
 
 	}
 
-	private void createPlayer(float x, float y) {
+	public void buildPlayer(float x, float y) {
 		BodyDef bodyDef = new BodyDef();
 		bodyDef.type = BodyDef.BodyType.DynamicBody;
 		bodyDef.position.set(x + 8, y + 8);
@@ -112,6 +106,7 @@ public class WorldBuilder {
 		entity.add(new StateComponent(PlayerComponent.IDLE_DOWN));
 		entity.add(new TextureComponent(textureRegion));
 		entity.add(new HealthComponent(500, 500));
+		entity.add(new HealthBarComponent(500, 500, stage));
 
 		AnimationComponent animationComponent = new AnimationComponent();
 		Animation<TextureRegion> animation;
@@ -142,52 +137,21 @@ public class WorldBuilder {
 
 		keyFrames.clear();
 
-		// move
-//		for (int i = 1; i < 3; i++) {
-//			keyFrames.add(new TextureRegion(actorAtlas.findRegion("Pacman"), i * 16, 0, 16, 16));
-//		}
 		animation = new Animation<TextureRegion>(0.1f, actorAtlas.findRegions("walkright"), PlayMode.LOOP);
 		animationComponent.animations.put(PlayerComponent.MOVE_RIGHT, animation);
 
-//		keyFrames.clear();
-
-//		for (int i = 3; i < 5; i++) {
-//			keyFrames.add(new TextureRegion(actorAtlas.findRegion("Pacman"), i * 16, 0, 16, 16));
-//		}
 		animation = new Animation<TextureRegion>(0.1f, actorAtlas.findRegions("walkleft"), PlayMode.LOOP);
 		animationComponent.animations.put(PlayerComponent.MOVE_LEFT, animation);
 
-//		keyFrames.clear();
-
-//		for (int i = 5; i < 7; i++) {
-//			keyFrames.add(new TextureRegion(actorAtlas.findRegion("Pacman"), i * 16, 0, 16, 16));
-//		}
 		animation = new Animation<TextureRegion>(0.1f, actorAtlas.findRegions("walkup"), PlayMode.LOOP);
 		animationComponent.animations.put(PlayerComponent.MOVE_UP, animation);
 
-//		keyFrames.clear();
-
-//		for (int i = 7; i < 9; i++) {
-//			keyFrames.add(new TextureRegion(actorAtlas.findRegion("Pacman"), i * 16, 0, 16, 16));
-//		}
 		animation = new Animation<TextureRegion>(0.1f, actorAtlas.findRegions("walkdown"), PlayMode.LOOP);
 		animationComponent.animations.put(PlayerComponent.MOVE_DOWN, animation);
-
-//		keyFrames.clear();
-
-//		for (int i = 0; i < 10; i++) {
-//			keyFrames.add(new TextureRegion(actorAtlas.findRegion("Pacman"), i * 16, 16, 16, 16));
-//		}
-//		keyFrames.add(new TextureRegion(actorAtlas.findRegion("Pacman"), 9 * 16, 0, 16, 16)); // invisible
-//		animation = new Animation(0.1f, keyFrames, Animation.PlayMode.NORMAL);
-//		animationComponent.animations.put(PlayerComponent.DIE, animation);
-//
-//		keyFrames.clear();
 
 		entity.add(animationComponent);
 
 		engine.addEntity(entity);
-//		body.setUserData(entity);
 	}
 
 }
