+++ core/src/io/alfami/components/HealthComponent.java
@@ -1,7 +1,6 @@
 package io.alfami.components;
 
 import com.badlogic.ashley.core.Component;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
 
 public class HealthComponent implements Component {
