--- core/src/legacy/io/alfami/behaviour/PlayerMovement.java
@@ -1,9 +1,0 @@
-package legacy.io.alfami.behaviour;
-/**
- * enum for different player directions
- * @author esadhrbatovic
- *
- */
-public enum PlayerMovement {
-	LEFT, RIGHT, UP, DOWN, NONE;
-}
